vtimer: FIXME: kvm: vmx: Sync the guest tsc_deadline
Since the VM set the tsc_deadline msr directly, the hypervisor needs to know
the value on every VM exit. (In fact, we can optimize to get it when
necesary, but leave it for now)

FIXME: see if we can read the virtual tsc_deadline at once instead of
using two instructions.

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>