vtimer: kvm: Schedule/cancel virt timer on vcpu load and put respectively
Different from hv timer, we need to touch vtimer on vcpu load and put.
It's because hv timer only runs when the VM is running so it doesn't
need to start/stop explicitely. We only take care of hv timer when the
VM is going to non-running state.

In contrast, we can run vtimer even if VM is not running(i.e. vmx
operation), and the hypervisor needs to explicitely start and stop the vtimer when the vcpu
is loaded/put.

TODO: rewrite the commit message

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>

Conflicts:
	arch/x86/kvm/vmx.c