ipi: alloc cpu_ir_table dynamically
Using cpu_ir_table alloced in kvm struture resulted in not correcly
mapping the gpa to hva in the host hypervisor. The GPA address range for
the one in kvm structure is 0x410000000000 while the kalloced mem is in
0x660000000, which is also the case for vmx_vcpu members.

The difference is that kvm structure is alloced by vzalloc() and vcpu is
alloced by kmem_cache_zalloc().(i.e. vmallc vs kmalloc)

<From L1>
[   10.897960] CPU IR Table in GPA: 0x410003e7b4f0
[   10.898447] CPU IR Table test in GPA: 0x410003e7b4d8
[   10.898971] CPU IR Table test P in GPA: 0x661249520

With the GPA in 0x41xx, the result of kmap() in the host hyp is not
starting with 0xffff and accessing the mapped memory resulted in
protection faults.

<From L0>
[   57.973682] ipi_opt_enable: 1
[   74.880424] general protection fault: 0000 [#1] SMP PTI
[   74.886260] CPU: 2 PID: 5381 Comm: qemu-system-x86 Not tainted 4.18.0-ipi-recursive-l0 #15
[   74.895482] Hardware name: Cisco Systems Inc UCSC-C220-M5SX/UCSC-C220-M5SX, BIOS C220M5.3.1.3d.0.0613181103 06/13/2018
[   74.907424] RIP: 0010:kvm_emulate_hypercall+0x373/0x6f0

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>