FIMXE: set the table address for each vcpu load
We only set the table address when creating vcpus. This is enough if the
table address is in VMCS. But since we are using hypercalls and the
virtual hardware keeps the table address in per L1 vcpu variables, the
virtual hardware can’t find the proper address if L2 vcpu is migrated to
another L1 vcpu.

As a temporary solution, set the table address in L1 whenever L2 vcpu is
loaded, which covers cpu migration.

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>