Do not clear shadow ir table
See the comments in the code.

This fixed getting wrong destination in L0. For example, L0 gets the
right mapped pi_desc addr, but when it comes to get the offset, the
table is reset and the pi_desc ends up with pointing a wrong addr. I
observed that ndst becomes 9, which is not even a range of # of vcpus
and L0 goes wrong from there.

This patch fixes the frequent problem, but L3 still hangs on boot or on
running hackbench sometimes. L0~L2 are live, but doesn't show any
messages.

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>