Sync shadow pi_desc to the real pi_desc on VM entry
VM will take a look at pi_desc, especially when it got 0xf1 wakeup irq.
We sync shadow pi_desc to the real pi_desc so that VM hypervisor can
wake up nVM based on the real pi_desc.

This fixes not waking up L3 VM when it is supposed to get IPI.

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>