Check if vtimer interrupt is for VM or nVM
So far, L1 hypervisor knows that vtimer interrupt is only for its VM
(i.e. L2 VM). But now L1 hypervisor needs to handle cases where L3 is
also using the vtimer. Check if the vtimer is for which VM. We always
set the vtimer context on L2 or L3 entry, so the vtimer interrupt is for
the running VM.

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>