Don't save vTSC context twice, which is wrong
We already saving the context set_vtimer_nvm_exit(), and doing it again
would mess up the context. For example, the first save is fine, but if
we cancel the vtimer and save THAT context again, then we just lose the
original context.

Don't save it twice, but just once.

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>