Move set_nvm_timer on exit/entry close to the is_guest_mode flag setting.
Since the hypervisor decides whether the virq is for L2 or L3 depending
on the current VM status, let's change the vtimer context as close as
possible to the guest mode flag change. This would avoid a wrong timer
irq delivery. There's still a chance for a short window, though.

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>