Restore the vtimer context after switching to VMCS02
Virtual timer sets the real timer (i.e. L0 soft timer) based on the
tsc_offset in the VMCS. Since the vtimer is first programmed in the
nested VM with tsc offset in VMCS02, the hypervisor also need to restore
the nVM timer context with the VMCS02. Otherwise, the hardware will set
the timer based on tsc_offset in vmcs01 wrongly.

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>