Dont setup timer stuff in vcpu_load and put while vmcs switch
In vcpu_load and put, there's no way to exactly know how to restore
timer context since it doesn't know if it's in the guest state or not.
For example in vmcs_switch.

Skip timer stuff in vmcs switch since we will handle it directly.

This patch resolves the issue that L3 (unit test) gets interrupts even
thought it didn't set it. That is set with the L2 apic timer context by
vcpu_load in vmcs switch.

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>