Move set_vtimer_nvm_exit_pre() to a correct place
This function is called in a weird place where the flag is still in
guest_mode, but the vcpu->arch.tsc_offset is already set to the VM's.
As a result, calling kvm_x86_ops->read_l1_tsc_offset() function returns
(l1 offset - vmcs12->tsc_offset) where it always is supposed to return
l1 offset regardless of guest_mode status. It's all about mismatching
between guest_mode status and vcpu->arch.tsc_offset.

Move the function after the flag change so that we at least are in the
consistent state for the read_l1_tsc_offset(). In fact, this function
doesn't have to be before vmcs change/flag change, so it's a perfect
place to be.

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>