nvmx: kvm: copy the guest activity state from vmcs12 to vmcs02 on L2 entry
Without this patch applied to L1 when running L3 with HLT trap disabled
in L1 and L2, L3 can't boot to the completion; it can't bring secondary
cpus and got error before getting the shell. Typical problem with not
setting the guest activity state is that when the state is set to active
where it should be halt, and on the entry to VM, the VM will execute
instructions while it should have waited in hlt state waiting for
events.

In general, it is obvious to take the guest activity state set by L1
when running L2, which was missing before and is fixed here.

The opposite is not always true though. When L1 disabled HLT trap, but
L0 did not, the running L2 would exit to L0 due to the HLT instruction
execution. L0 observes that the guest activity state in vmcs02 is still
0 (active), but when entering L1 it needs to set it to 1(hlt) in vmcs12
so that L1 can observe HLT state of L2. We don't touch this part in this
patch.

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>