ipi: set cpu_ir_table correctly
vmx_vcpu_load() can be called for either L1 or L2 context, e.g. when
switching vmcses. Therefore, set cpu_ir_table correctly for L1.

We also need to set one for L2 as well, but we don't. It is fine
probably because L0 keeps the previous value. For example, when a vcpu
having L2 context is loaded, it doesn't set cpu_ir_table for L2, but the
L0 already remembers cpu_ir_table for L2 from the previous execution.

When switchng vmcses, the hypervisor sets cpu_ir_table for L2 explicitly
in nested_vmx_run().

TODO: this doesn't look good, but will be revised when the cpu_ir_table
is passed through vmcs later

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>