ipi: set pir for L2 correctly
There could a race condition with the previous code, which I don't
remember now. This is the comment on the day before SOSP 19 submission.

race condition setting PI disc: L1 is about to make L3 idle, but L0 on
the other cpu sends with 0xf0 nv. After setting ON bit correctly, things
are fine.

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>