Don't try to wake up vcpu in the context of schedule()
When running schedule(), e.g. to wait for the timer interrupt, we (L1)
cancels all hardware timers and set up sw timers. When  the timer is
expired at that moment, we tried to wake up this vcpu again which
requires a lock that already held by schedule(). As a result, L1 gets
stuck.

As a walkaround, we don't wake up the vcpu in the context of schedule()
expecting that L1 hypervisor would get an interrupt for the expired sw
timer after schedule() is done (and the interrupt is enabled again).

*TODO*: maybe the right solution is to switch between hw and sw timer in
pre_block and post_block like preemption timer management.

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>